requirements.txt
chromadb
sentence_transformers
langchain_huggingface
faiss_cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
pymupdf
arxiv
wikipedia
validators==0.28.1
youtube_transcript_api
